use same sidebar for all three pod views
